+++ README.md
@@ -1,5 +1,7 @@
 # DeepEmotion
-# Team Members
+# Team Members:
+
+Fan Lu, Xi Yu 
 # Dataset Description
 The data consists of 48x48 pixel grayscale images of faces. The faces have been automatically registered so that the face is more or less centered and occupies about the same amount of space in each image. The task is to categorize each face based on the emotion shown in the facial expression in to one of seven categories (0=Angry, 1=Disgust, 2=Fear, 3=Happy, 4=Sad, 5=Surprise, 6=Neutral).
