+++ capsuleNet_keras/code/CNN_fer2013.py
@@ -1,0 +1,87 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Apr 13 22:21:21 2018
+
+@author: Xi Yu
+"""
+
+import numpy as np
+import os
+import csv
+import matplotlib.pyplot as plt
+import keras
+from keras.datasets import mnist
+from keras.models import Sequential
+from keras.layers import Dense, Dropout, Flatten
+from keras.layers import Conv2D, MaxPooling2D
+from keras import backend as K
+#%%
+
+def loadfer2013():
+    
+    tmp = np.loadtxt("fer2013.csv", dtype=np.str, delimiter=",")
+    reader = tmp[1:,1]
+    TRAIN_END_POINT = len(reader)
+    train_data_x = np.zeros([TRAIN_END_POINT, 48, 48], dtype="uint8")
+    for k, data in enumerate(reader):
+       pixels_formated = [int(a) for a in data.split(" ")] 
+       pixels_in_picture_format = np.reshape(pixels_formated, [48, 48])
+       train_data_x[k, :, :] = pixels_in_picture_format
+    label = tmp[1:,0].astype(np.uint8)
+    return train_data_x, label
+
+data, label = loadfer2013()
+#%%
+x_train = data[0:28700]
+y_train = label[0:28700]
+x_validation = data[28709:32298]
+y_validation = label[28709:32298]
+x_test = data[32298:]
+y_test = label[32298:]
+
+img_rows, img_cols = 48, 48
+batch_size = 128
+num_classes = 7
+epochs = 3
+
+if K.image_data_format() == 'channels_first':
+    x_train = x_train.reshape(x_train.shape[0], 1, img_rows, img_cols)
+    x_test = x_test.reshape(x_test.shape[0], 1, img_rows, img_cols)
+    input_shape = (1, img_rows, img_cols)
+else:
+    x_train = x_train.reshape(x_train.shape[0], img_rows, img_cols, 1)
+    x_test = x_test.reshape(x_test.shape[0], img_rows, img_cols, 1)
+    input_shape = (img_rows, img_cols, 1)
+
+
+
+x_train = x_train.astype('float32')
+x_validation = x_validation.astype('float32')
+x_test = x_test.astype('float32')
+x_train /= 255
+x_validation /=255
+x_test /= 255
+print('x_train shape:', x_train.shape)
+print(x_train.shape[0], 'train samples')
+print(x_test.shape[0], 'test samples')
+
+y_train = keras.utils.to_categorical(y_train, num_classes)
+y_test = keras.utils.to_categorical(y_test, num_classes)
+#%%
+model = Sequential()
+model.add(Conv2D(32, kernel_size=(3, 3), activation='relu',input_shape=input_shape))
+model.add(Conv2D(64, (3, 3), activation='relu'))
+model.add(MaxPooling2D(pool_size=(2, 2)))
+model.add(Dropout(0.25))
+model.add(Flatten())
+model.add(Dense(128, activation='relu'))
+model.add(Dropout(0.5))
+model.add(Dense(num_classes, activation='softmax'))
+model.summary()
+
+model.compile(loss=keras.losses.categorical_crossentropy,optimizer=keras.optimizers.Adadelta(),metrics=['accuracy'])
+
+model.fit(x_train, y_train,batch_size=batch_size,epochs=epochs,verbose=1, validation_data=(x_validation, y_validation))
+score = model.evaluate(x_test, y_test, verbose=0)
+print('Test loss:', score[0])
+print('Test accuracy:', score[1])
