+++ capsuleNet_keras/screenshots/filter 4x4.png
